Add the ability to assign on call users to an approver group in Gitlab.
You configure deputize to map the users in a PagerDuty oncall schedule to a group in Gitlab. This group can be assigned as an approver of merge requests for any/all of your Gitlab projects.

This can be useful to ensure infrastructure or other changes have a viable approver available during oncall incidents, as opposed to pulling in another person just to approve changes